wasm: Fixes and more bindings.
--

Add this._doc to PDFObject, PDFPage, PDFAnnot.

We need this for passing to PDFOBJ(doc, obj) in order to create
PDF dictionaries and arrays from native Javascript objects.

--

Compile with WASM_ASYNC_COMPILATION=0 to avoid the need for "await" or
a "ready" callback.

--

Lots of new functions and methods exposed.
Lots of fixes to existing bindings.